requirements.txt
beautifulsoup4==4.11.1
colorama==0.4.4
hydra-core==1.2.0
librosa==0.8.0
omegaconf==2.2.3
pygame==2.1.2
requests==2.25.1
SoundFile==0.10.3.post1
whisper==1.0
